[BE] REFACTOR: DTO 파일 위치 변경
