+++ backend/src/main/java/org/ftclub/cabinet/presentation/dto/PresentationCommentRequestDto.java
@@ -1,4 +1,4 @@
-package org.ftclub.cabinet.dto;
+package org.ftclub.cabinet.presentation.dto;
 
 import javax.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
