+++ backend/src/main/java/org/ftclub/cabinet/presentation/dto/PresentationCommentServiceCreationDto.java
@@ -1,18 +1,16 @@
-package org.ftclub.cabinet.dto;
+package org.ftclub.cabinet.presentation.dto;
+
 
 import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 
-@EqualsAndHashCode
 @Getter
 @ToString
 @AllArgsConstructor
-public class PresentationCommentServiceUpdateDto {
+public class PresentationCommentServiceCreationDto {
 
 	private final Long userId;
 	private final Long presentationId;
-	private final Long commentId;
 	private final String commentDetail;
 }
