+++ backend/src/main/java/org/ftclub/cabinet/presentation/service/PresentationCommentService.java
@@ -3,9 +3,9 @@ package org.ftclub.cabinet.presentation.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import org.ftclub.cabinet.dto.PresentationCommentResponseDto;
-import org.ftclub.cabinet.dto.PresentationCommentServiceCreationDto;
-import org.ftclub.cabinet.dto.PresentationCommentServiceUpdateDto;
+import org.ftclub.cabinet.presentation.dto.PresentationCommentResponseDto;
+import org.ftclub.cabinet.presentation.dto.PresentationCommentServiceCreationDto;
+import org.ftclub.cabinet.presentation.dto.PresentationCommentServiceUpdateDto;
 import org.ftclub.cabinet.exception.ExceptionStatus;
 import org.ftclub.cabinet.presentation.domain.Presentation;
 import org.ftclub.cabinet.presentation.domain.PresentationComment;
