+++ backend/src/test/java/org/ftclub/cabinet/presentation/controller/PresentationCommentControllerTest.java
@@ -18,15 +18,15 @@ import java.util.Arrays;
 import java.util.List;
 import org.ftclub.cabinet.alarm.discord.DiscordWebHookMessenger;
 import org.ftclub.cabinet.auth.service.AuthPolicyService;
-import org.ftclub.cabinet.dto.PresentationCommentServiceCreationDto;
-import org.ftclub.cabinet.dto.PresentationCommentServiceUpdateDto;
+import org.ftclub.cabinet.presentation.dto.PresentationCommentServiceCreationDto;
+import org.ftclub.cabinet.presentation.dto.PresentationCommentServiceUpdateDto;
 import org.ftclub.cabinet.dto.UserInfoDto;
 import org.ftclub.cabinet.jwt.service.JwtService;
 import org.ftclub.cabinet.security.exception.SecurityExceptionHandlerManager;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.time.LocalDateTime;
-import org.ftclub.cabinet.dto.PresentationCommentRequestDto;
-import org.ftclub.cabinet.dto.PresentationCommentResponseDto;
+import org.ftclub.cabinet.presentation.dto.PresentationCommentRequestDto;
+import org.ftclub.cabinet.presentation.dto.PresentationCommentResponseDto;
 import org.ftclub.cabinet.presentation.service.PresentationCommentService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
